Fix lsu LQ lock, and much refractoring, got 4 core to boot linux and run user space terminal in sim
